requirements.txt
scapy
python-pcapng